chore(auto-place): move common feature to diagram-js
https://github.com/bpmn-io/dmn-js/issues/470
