--- lib/features/auto-place/AutoPlaceSelectionBehavior.js
@@ -1,18 +1,0 @@
-/**
- * Select element after auto placement.
- *
- * @param {EventBus} eventBus
- * @param {Selection} selection
- */
-export default function AutoPlaceSelectionBehavior(eventBus, selection) {
-
-  eventBus.on('autoPlace.end', 500, function(e) {
-    selection.select(e.shape);
-  });
-
-}
-
-AutoPlaceSelectionBehavior.$inject = [
-  'eventBus',
-  'selection'
-];
