+++ lib/features/auto-place/index.js
@@ -1,8 +1,9 @@
-import AutoPlace from './AutoPlace';
-import AutoPlaceSelectionBehavior from './AutoPlaceSelectionBehavior';
+import AutoPlaceModule from 'diagram-js/lib/features/auto-place';
+
+import BpmnAutoPlace from './BpmnAutoPlace';
 
 export default {
-  __init__: [ 'autoPlaceSelectionBehavior' ],
-  autoPlace: [ 'type', AutoPlace ],
-  autoPlaceSelectionBehavior: [ 'type', AutoPlaceSelectionBehavior ]
+  __depends__: [ AutoPlaceModule ],
+  __init__: [ 'bpmnAutoPlace' ],
+  bpmnAutoPlace: [ 'type', BpmnAutoPlace ]
 };
