+++ lib/features/grid-snapping/behavior/AutoPlaceBehavior.js
@@ -1,11 +1,13 @@
-import { getNewShapePosition } from '../../auto-place/AutoPlaceUtil';
+import { getNewShapePosition } from '../../auto-place/BpmnAutoPlaceUtil';
 
 import { getMid } from 'diagram-js/lib/layout/LayoutUtil';
 import { is } from '../../../util/ModelUtil';
 
+var HIGH_PRIORITY = 2000;
+
 
 export default function AutoPlaceBehavior(eventBus, gridSnapping) {
-  eventBus.on('autoPlace', function(context) {
+  eventBus.on('autoPlace', HIGH_PRIORITY, function(context) {
     var source = context.source,
         sourceMid = getMid(source),
         shape = context.shape;
