+++ test/spec/features/auto-place/BpmnAutoPlace.boundary-events.bpmn
