+++ test/spec/features/auto-place/BpmnAutoPlace.bpmn
