+++ test/spec/features/auto-place/BpmnAutoPlace.multi-connection.bpmn
